+++ test/test_xmlmap/test_ead.py
@@ -14,10 +14,15 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+from __future__ import unicode_literals
 import unittest
-from unittest2 import skipIf
+try:
+    from unittest import skipIf
+except ImportError:
+    from unittest2 import skipIf
 import os
 
+from eulxml.utils.compat import u
 from eulxml.xmlmap import load_xmlobject_from_file, load_xmlobject_from_string
 from eulxml.xmlmap import eadmap
 
@@ -36,19 +41,19 @@ class TestEad(unittest.TestCase):
         self.assert_(isinstance(self.ead, eadmap.EncodedArchivalDescription))
 
     def test_basic_fields(self):
-        self.assertEqual(unicode(self.ead.title), "Seamus Heaney collection, 1972-1997")
-        self.assertEqual(unicode(self.ead.eadid), u'heaney653')
+        self.assertEqual(u(self.ead.title), "Seamus Heaney collection, 1972-1997")
+        self.assertEqual(u(self.ead.eadid), 'heaney653')
         self.assertEqual(self.ead.id, "heaney653-011")
         self.assertEqual(self.ead.author, "Manuscript, Archives, and Rare Book Library, Emory University")
         # whitespace makes fields with tags a bit messier...
-        self.assert_("Seamus Heaney collection," in unicode(self.ead.unittitle))
-        self.assert_("1972-2005" in unicode(self.ead.unittitle))
+        self.assert_("Seamus Heaney collection," in u(self.ead.unittitle))
+        self.assert_("1972-2005" in u(self.ead.unittitle))
         # several different extents in the physical description;
         # FIXME: all smashed together
         self.assert_("1 linear ft." in self.ead.physical_desc)
         self.assert_("(3 boxes)" in self.ead.physical_desc)
         self.assert_("12 oversized papers (OP)" in self.ead.physical_desc)
-        self.assert_("materials relating to Irish poet Seamus Heaney" in unicode(self.ead.abstract))
+        self.assert_("materials relating to Irish poet Seamus Heaney" in u(self.ead.abstract))
 
     @proxy_required
     def test_validation(self):
@@ -70,7 +75,7 @@ class TestEad(unittest.TestCase):
         self.assertEqual("Heaney, Seamus, 1939-", ad.origination)
         self.assert_(isinstance(ad.unitid, eadmap.Unitid))
         self.assertEqual("Manuscript Collection No.653", ad.unitid.value)
-        self.assertEqual("Manuscript Collection No.653", unicode(ad.unitid))
+        self.assertEqual("Manuscript Collection No.653", u(ad.unitid))
         self.assertEqual('US', ad.unitid.country_code)
         self.assertEqual('geu-s', ad.unitid.repository_code)
         self.assertEqual(653, ad.unitid.identifier)
@@ -80,48 +85,48 @@ class TestEad(unittest.TestCase):
         self.assertEqual("Materials entirely in English.", ad.langmaterial)
         self.assertEqual("In the Archives.", ad.location)
         self.assert_(isinstance(ad.access_restriction, eadmap.Section))
-        self.assertEqual("Restrictions on access", unicode(ad.access_restriction.head))
-        self.assert_("Special restrictions apply" in unicode(ad.access_restriction.content[0]))
+        self.assertEqual("Restrictions on access", u(ad.access_restriction.head))
+        self.assert_("Special restrictions apply" in u(ad.access_restriction.content[0]))
         self.assert_(isinstance(ad.use_restriction, eadmap.Section))
-        self.assertEqual("Terms Governing Use and Reproduction", unicode(ad.use_restriction.head))
-        self.assert_("limitations noted in departmental policies" in unicode(ad.use_restriction.content[0]))
+        self.assertEqual("Terms Governing Use and Reproduction", u(ad.use_restriction.head))
+        self.assert_("limitations noted in departmental policies" in u(ad.use_restriction.content[0]))
         self.assert_(isinstance(ad.alternate_form, eadmap.Section))
-        self.assertEqual("Publication Note", unicode(ad.alternate_form.head))
-        self.assert_("Published in" in unicode(ad.alternate_form.content[0]))
+        self.assertEqual("Publication Note", u(ad.alternate_form.head))
+        self.assert_("Published in" in u(ad.alternate_form.content[0]))
         self.assert_(isinstance(ad.originals_location, eadmap.Section))
-        self.assertEqual("Location of Originals", unicode(ad.originals_location.head))
-        self.assert_("Suppressed chapter" in unicode(ad.originals_location.content[0]))
+        self.assertEqual("Location of Originals", u(ad.originals_location.head))
+        self.assert_("Suppressed chapter" in u(ad.originals_location.content[0]))
         self.assert_(isinstance(ad.related_material, eadmap.Section))
-        self.assertEqual("Related Materials in This Repository", unicode(ad.related_material.head))
-        self.assert_("part of MSS" in unicode(ad.related_material.content[0]))
+        self.assertEqual("Related Materials in This Repository", u(ad.related_material.head))
+        self.assert_("part of MSS" in u(ad.related_material.content[0]))
         self.assert_(isinstance(ad.separated_material, eadmap.Section))
-        self.assertEqual("Related Materials in This Repository", unicode(ad.separated_material.head))
-        self.assert_("Ciaran Carson papers, Peter Fallon" in unicode(ad.separated_material.content[0]))
+        self.assertEqual("Related Materials in This Repository", u(ad.separated_material.head))
+        self.assert_("Ciaran Carson papers, Peter Fallon" in u(ad.separated_material.content[0]))
         self.assert_(isinstance(ad.acquisition_info, eadmap.Section))
-        self.assertEqual("Source", unicode(ad.acquisition_info.head))
-        self.assert_("Collection assembled from various sources." in unicode(ad.acquisition_info.content[0]))
+        self.assertEqual("Source", u(ad.acquisition_info.head))
+        self.assert_("Collection assembled from various sources." in u(ad.acquisition_info.content[0]))
         self.assert_(isinstance(ad.custodial_history, eadmap.Section))
-        self.assertEqual("Custodial History", unicode(ad.custodial_history.head))
-        self.assert_("Originally received as part of" in unicode(ad.custodial_history.content[0]))
+        self.assertEqual("Custodial History", u(ad.custodial_history.head))
+        self.assert_("Originally received as part of" in u(ad.custodial_history.content[0]))
         self.assert_(isinstance(ad.preferred_citation, eadmap.Section))
-        self.assertEqual("Citation", unicode(ad.preferred_citation.head))
-        self.assert_("[after identification of item(s)" in unicode(ad.preferred_citation.content[0]))
+        self.assertEqual("Citation", u(ad.preferred_citation.head))
+        self.assert_("[after identification of item(s)" in u(ad.preferred_citation.content[0]))
         self.assert_(isinstance(ad.biography_history, eadmap.Section))
-        self.assertEqual("Biographical Note", unicode(ad.biography_history.head))
-        self.assert_("born on April 13" in unicode(ad.biography_history.content[0]))
-        self.assert_("While at St. Joseph's" in unicode(ad.biography_history.content[1]))
+        self.assertEqual("Biographical Note", u(ad.biography_history.head))
+        self.assert_("born on April 13" in u(ad.biography_history.content[0]))
+        self.assert_("While at St. Joseph's" in u(ad.biography_history.content[1]))
         self.assert_(isinstance(ad.bibliography, eadmap.Section))
-        self.assertEqual("Publication Note", unicode(ad.bibliography.head))
-        self.assert_("Susan Jenkins Brown" in unicode(ad.bibliography.content[0]))
+        self.assertEqual("Publication Note", u(ad.bibliography.head))
+        self.assert_("Susan Jenkins Brown" in u(ad.bibliography.content[0]))
         self.assert_(isinstance(ad.scope_content, eadmap.Section))
-        self.assertEqual("Scope and Content Note", unicode(ad.scope_content.head))
-        self.assert_("consists of materials relating" in unicode(ad.scope_content.content[0]))
+        self.assertEqual("Scope and Content Note", u(ad.scope_content.head))
+        self.assert_("consists of materials relating" in u(ad.scope_content.content[0]))
         self.assert_(isinstance(ad.arrangement, eadmap.Section))
-        self.assertEqual("Arrangement Note", unicode(ad.arrangement.head))
-        self.assert_("five series" in unicode(ad.arrangement.content[0]))
+        self.assertEqual("Arrangement Note", u(ad.arrangement.head))
+        self.assert_("five series" in u(ad.arrangement.content[0]))
         self.assert_(isinstance(ad.other, eadmap.Section))
-        self.assertEqual("Finding Aid Note", unicode(ad.other.head))
-        self.assert_("Index to selected correspondents" in unicode(ad.other.content[0]))
+        self.assertEqual("Finding Aid Note", u(ad.other.head))
+        self.assert_("Index to selected correspondents" in u(ad.other.content[0]))
 
     def test_index_indexentry(self):
         ad = self.ead.archdesc
@@ -129,31 +134,31 @@ class TestEad(unittest.TestCase):
         self.assertEqual(2, len(ad.index))
         index = ad.index[0]
         self.assert_(isinstance(index, eadmap.Index))
-        self.assertEqual('Index of Selected Correspondents', unicode(index.head))
+        self.assertEqual('Index of Selected Correspondents', u(index.head))
         self.assertEqual('index1', index.id)
-        self.assert_('relates to the correspondence in Series 1' in unicode(index.note.content[0]))
+        self.assert_('relates to the correspondence in Series 1' in u(index.note.content[0]))
         self.assertEqual(2, len(index.entry))
         self.assert_(isinstance(index.entry[0], eadmap.IndexEntry))
-        self.assertEqual(u'Batten, Guinn', unicode(index.entry[0].name))
+        self.assertEqual('Batten, Guinn', u(index.entry[0].name))
         self.assert_(isinstance(index.entry[0].ptrgroup, eadmap.PointerGroup))
         self.assertEqual(3, len(index.entry[0].ptrgroup.ref))
         self.assert_(isinstance(index.entry[0].ptrgroup.ref[0], eadmap.Reference))
         self.assertEqual('simple', index.entry[0].ptrgroup.ref[0].type)
-        self.assert_(u'1995 July' in unicode(index.entry[0].ptrgroup.ref[0].value))
-        self.assertEqual(u'Belton, Neil', unicode(index.entry[1].name))
-        self.assert_(u'1993 November 3' in unicode(index.entry[1].ptrgroup.ref[-1].value))
+        self.assert_('1995 July' in u(index.entry[0].ptrgroup.ref[0].value))
+        self.assertEqual('Belton, Neil', u(index.entry[1].name))
+        self.assert_('1993 November 3' in u(index.entry[1].ptrgroup.ref[-1].value))
 
         # multiple indexes
         self.assert_(isinstance(ad.index[1], eadmap.Index))
-        self.assertEqual("Second Index", unicode(ad.index[1].head))
+        self.assertEqual("Second Index", u(ad.index[1].head))
         self.assertEqual("index2", ad.index[1].id)
 
     def test_ControlledAccessHeadings(self):
         ca = self.ead.archdesc.controlaccess
         self.assert_(isinstance(ca, eadmap.ControlledAccessHeadings))
-        self.assertEqual("Selected Search Terms", unicode(ca.head))
+        self.assertEqual("Selected Search Terms", u(ca.head))
         self.assert_(isinstance(ca.controlaccess[0], eadmap.ControlledAccessHeadings))
-        self.assertEqual("Personal Names", unicode(ca.controlaccess[0].head))
+        self.assertEqual("Personal Names", u(ca.controlaccess[0].head))
         self.assert_(isinstance(ca.controlaccess[0].person_name[0], eadmap.Heading))
         self.assertEqual("Barker, Sebastian.", ca.controlaccess[0].person_name[0].value)
         self.assert_(isinstance(ca.controlaccess[0].family_name[0], eadmap.Heading))
@@ -185,19 +190,19 @@ class TestEad(unittest.TestCase):
         dsc = self.ead.dsc
         self.assert_(isinstance(dsc, eadmap.SubordinateComponents))
         self.assertEqual("combined", dsc.type)
-        self.assertEqual("Description of Series", unicode(dsc.head))
+        self.assertEqual("Description of Series", u(dsc.head))
         # c01 - series
         self.assert_(isinstance(dsc.c[0], eadmap.Component))
         self.assertEqual("series", dsc.c[0].level)
         self.assert_(isinstance(dsc.c[0].did, eadmap.DescriptiveIdentification))
         self.assertEqual("Series 1", dsc.c[0].did.unitid.value)
-        self.assertEqual("Writings by Seamus Heaney", unicode(dsc.c[0].did.unittitle))
+        self.assertEqual("Writings by Seamus Heaney", u(dsc.c[0].did.unittitle))
         self.assertEqual("Box 1: folders 1-12", dsc.c[0].did.physdesc)
         # c02 - file
         self.assert_(isinstance(dsc.c[0].c[0], eadmap.Component))
         self.assertEqual("file", dsc.c[0].c[0].level)
         self.assert_(isinstance(dsc.c[0].c[0].did, eadmap.DescriptiveIdentification))
-        self.assert_("holograph manuscript" in unicode(dsc.c[0].c[0].did.unittitle))
+        self.assert_("holograph manuscript" in u(dsc.c[0].c[0].did.unittitle))
         self.assertEqual("box", dsc.c[0].c[0].did.container[0].type)
         self.assertEqual("1", dsc.c[0].c[0].did.container[0].value)
         self.assertEqual("folder", dsc.c[0].c[0].did.container[1].type)
@@ -210,7 +215,7 @@ class TestEad(unittest.TestCase):
         self.assertTrue(dsc.c[1].hasSubseries())
         # access c03 level item
         self.assertEqual("file", dsc.c[1].c[0].c[0].level)
-        self.assert_("Hilary Boyle" in unicode(dsc.c[1].c[0].c[0].did.unittitle))
+        self.assert_("Hilary Boyle" in u(dsc.c[1].c[0].c[0].did.unittitle))
 
     def test_SubordinateComponents_noseries(self):
         # simple finding aid with no series but only a container list
@@ -236,14 +241,14 @@ class TestEad(unittest.TestCase):
         self.assertEqual("eng", profiledesc.language_codes[0])
         # profile creation date
         self.assert_(isinstance(profiledesc.date, eadmap.DateField))
-        self.assertEqual(u'May 5, 2005', unicode(profiledesc.date))
+        self.assertEqual('May 5, 2005', u(profiledesc.date))
         self.assertEqual('2005-05-05', profiledesc.date.normalized)
 
     def test_DateField(self):
         date = self.ead.file_desc.publication.date
         self.assert_(isinstance(date, eadmap.DateField))
         self.assertEqual("2005-05-05", date.normalized)
-        self.assertEqual("May 5, 2005", unicode(date))
+        self.assertEqual("May 5, 2005", u(date))
         self.assertEqual("ce", date.era)
         self.assertEqual("gregorian", date.calendar)
 
@@ -256,11 +261,11 @@ class TestEad(unittest.TestCase):
         self.assert_(isinstance(title, eadmap.UnitTitle))
         self.assert_(isinstance(title.short, eadmap.UnitTitle))
         self.assert_(isinstance(title.unitdate, eadmap.DateField))
-        self.assertEqual(u'1972-2005', unicode(title.unitdate))
+        self.assertEqual('1972-2005', u(title.unitdate))
         # short title
-        self.assertEqual(u'Seamus Heaney collection,', unicode(title.short))
-        self.assertEqual(u'Writings by Seamus Heaney',
-                         unicode(self.ead.dsc.c[0].did.unittitle.short))
+        self.assertEqual('Seamus Heaney collection,', u(title.short))
+        self.assertEqual('Writings by Seamus Heaney',
+                         u(self.ead.dsc.c[0].did.unittitle.short))
 
     def test_DigitalArchivalObject(self):
         # test dao inserted in did of first container-level item, series 1
@@ -270,4 +275,3 @@ class TestEad(unittest.TestCase):
         self.assertEqual('id1234', dao.id)
         self.assertEqual('Digitized Content', dao.title)
         self.assertEqual('http://some.url/item/id1234', dao.href)
-