+++ test/test_xmlmap/test_premis.py
@@ -14,8 +14,12 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+from __future__ import unicode_literals
 import unittest
-from unittest2 import skipIf
+try:
+    from unittest import skipIf
+except ImportError:
+    from unittest2 import skipIf
 import os
 
 from eulxml.xmlmap import load_xmlobject_from_file